Add files via upload
I still have to do some work to have it in the right stile and make sure that all the code follows the PEP 8 format but the problem with the function arguments being tuples has been dealt with. I think it could still use some work but I will look at it more in the morning.